Разделение ChessmenMovement.java на два класса: VerticalAndHorizontalMovement и DiagonalMovement
