+++ src/main/java/org/example/chess/Rook.java
@@ -1,6 +1,6 @@
 package org.example.chess;
 
-import org.example.movement.ChessmenMovement;
+import org.example.movement.VerticalAndHorizontalMovement;
 
 import java.util.List;
 
@@ -9,17 +9,9 @@ import java.util.List;
  */
 public class Rook implements Chessmen {
     /**
-     * Экземпляр класса chessmenMovement
+     * Экземпляр класса VerticalAndHorizontalMovement
      */
-    private final ChessmenMovement chessmenMovement = new ChessmenMovement();
-    /**
-     * Минимальная размерность игрового поля
-     */
-    private final static int MIN_SIDE_VALUE = 0;
-    /**
-     * Максимальная размерность игрового поля
-     */
-    private final static int MAX_SIDE_VALUE = 7;
+    private final VerticalAndHorizontalMovement allVerticalAndHorizontal = new VerticalAndHorizontalMovement();
 
     @Override
     public boolean checkMove(PositionOnBoard start, PositionOnBoard finish, byte[][] board) {
@@ -27,7 +19,7 @@ public class Rook implements Chessmen {
                 isPositionEnemy(start.row(), start.column(), finish.row(), finish.column(), board)) {
             if ((Math.abs(start.row() - finish.row()) == 0 ^
                     Math.abs(start.column() - finish.column()) == 0) &&
-                    chessmenMovement.isWayFree(start, finish, board)) {
+                    allVerticalAndHorizontal.isWayFree(start, finish, board)) {
                 return true;
             }
         }
@@ -51,7 +43,7 @@ public class Rook implements Chessmen {
 
     @Override
     public List<PositionOnBoard> allPossibleMoves(PositionOnBoard start, byte[][] board) {
-        List<PositionOnBoard> possibleMoves = chessmenMovement.allVerticalAndHorizontalMoves(start, board);
+        List<PositionOnBoard> possibleMoves = allVerticalAndHorizontal.allPossibleMoves(start, board);
         return possibleMoves;
     }
