+++ src/main/java/org/example/movement/Movement.java
@@ -1,0 +1,12 @@
+package org.example.movement;
+
+import org.example.chess.PositionOnBoard;
+
+import java.util.List;
+
+public interface Movement {
+    /**
+     * Вычисляет список доступных ходов
+     */
+    List<PositionOnBoard> allPossibleMoves(PositionOnBoard start, byte[][] board);
+}
